Added a limit to the am_layout draggable
Saving here, about to scrap the pm_layout, add a horizontal ScrollView to am_layout, add 24 time slots to it, then preserve the limited bar! Actually may be more simple than what I had in mind. As long as the ScrollView will actually scroll as a user drags, it should be okay.